capable to add geofence if configure the location mode to "High accuracy" From System Setting > Location > Mode

+++ app/src/main/java/btremote999/geofencetest/data/MyCircle.java
@@ -3,5 +3,5 @@ package btremote999.geofencetest.data;
 import com.google.gson.annotations.Expose;
 
 public class MyCircle extends MyArea {
-    @Expose public Double radius; // in meters
+    @Expose public float radius; // in meters
 }
